Create vocabulary dto

+++ src/main/java/com/learnenglish/studentsvocabulary/service/UserService.java
@@ -6,10 +6,10 @@ import com.learnenglish.studentsvocabulary.model.Vocabulary;
 import java.util.Set;
 
 public interface UserService {
-    User find(int id);
-    void delete(int id);
-    void addVocabulary(int id, Vocabulary vocabulary);
-    void detachVocabulary(int id, Vocabulary vocabulary);
+    User find(Long id);
+    void delete(Long id);
+    void addVocabulary(Long id, Vocabulary vocabulary);
+    void detachVocabulary(Long id, Vocabulary vocabulary);
     User create(User user);
-    Set<Vocabulary> getVocabularies(int id);
+    Set<Vocabulary> getVocabularies(Long id);
 }
